text_generators.sh: add option for loading uEnv.txt file
uEnv.txt can be stored on the boot partition and if present, is imported
to the u-boot environment. Additionally, if the uEnv.txt is present,
uenvcmd environment variable is executed as a command. This allows text
mode editing of boot parameters by directly writing to the uEnv.txt,
without need to re-compile boot scripts for every change.

This support is implemented by modifying the existing boot scripts only.

Signed-off-by: Tero Kristo <tero.kristo@unikie.com>