setup.sh: add support for updating filesystem image contents only
By default, bootfs/rootfs/domufs contents are completely wiped out with
rm -rf when installing new image. In some cases the user may want to
preserve the existing contents (e.g. user scripts / tools etc.) and a
new option is provided for this purpose: FS_UPDATE_ONLY. If this option
is used, the wipe portion is omitted and the new image is installed over
the existing one.

Signed-off-by: Tero Kristo <tero.kristo@unikie.com>
Change-Id: I0da9b2b7d8b04e34032b3e25a8927d72f510d0b4