setup.sh: add new option SUB_PLATFORM for x86
x86 platform differs between Intel and AMD, namely the KVM hypervisor
will fail with wrong config, and current setup does not work on AMD
processors. The fix is to swap the kernel module from CONFIG_KVM_INTEL
to CONFIG_KVM_AMD, and with this KVM will also work. To handle this
properly, add a new global config option to the build, SUB_PLATFORM,
which can take either intel or amd as parameter.

Jira-ID: SP-384
Signed-off-by: Tero Kristo <tero.kristo@unikie.com>
Change-Id: I477ce359ebe5759676d2e316298960e13aa38fac