configs: secure: raspi4: fix GCC plugin support
GCC plugins appear to cause problems with GCC 9.x on ARM64 targets. The
target device fails to boot with this combination, and the workaround is
to build the kernel with GCC 10.x. However, XEN tools fail to build with
GCC 10.x due to uninitialized variable warning which is interpreted as
an error with the newer compiler.

To workaround this bunch of issues, we do two things:
- disable GCC plugins in xen build,
- update the GCC version on KVM build and disable XEN tools.

Jira-Id: SP-422
Signed-off-by: Tero Kristo <tero.kristo@unikie.com>
Change-Id: I8d1ed76edaacdda29c69998214c17ca2158211ad