Fix ssh support for new build setup
Generate the RSA keys for SSH access, and copy these to the rootfs
image. Also, modify the SSH access script to autodetect target IP
address via the bridge, and use this automatically when connecting.

With this change, one can connect to admin VM with:
./setup.sh ssh_dut vm_admin

Same works for conn VM once it is up.

Also, fix the wrong bridge ID for br_conn while here.

Jira-Id: SP-483
Signed-off-by: Tero Kristo <tero.kristo@unikie.com>
Change-Id: Ie210e8a42159c0f5a4a4a39b1503f8e0d810d3d0