Add Secure-OS VM support
Initially secure-OS VM image consists of a hardened linux kernel, and
slightly customized rootfs (the main difference is inclusion of docker.)
The build system is also modified to build the image by default, and we
also fix some of the admin VM features to boot up all the VM images
properly. For now, secure-OS VM is connected to the same network bridge
as the connection VM, but this should be changed to go via connection VM
once support is ready. This change would be transparent to the VM
itself.

Jira-Id: SP-484
Signed-off-by: Tero Kristo <tero.kristo@unikie.com>
Change-Id: Ia3c2633156ab41a9d908cdc4832499ce578d08ee