text_generators: convert the output image files back to simple format
With the image name conversion, e.g. the rootfs image for the connection
VM changed to something like br_conn_x86_qemu.ext2, which the boot
script is no longer able to find. Modify the makefile to ditch these
extra portions out of the file so that the boot script can continue
working.

Jira-Id: SP-497
Signed-off-by: Tero Kristo <tero.kristo@unikie.com>
Change-Id: I9e948f8af428a144fbb33b902a5cfbccc7acd47d