Generate kernel defconfigs when required
Defconfigs will be regenerated from fragments, if there are changes.

Atm any changes to any fragments will trigger defconfig regeneration,
even if the fragmets are not used in a specific vm, but this should be
an advantage over the former way of handling things anyway.

Now defconfig generation uses the buildroot copy of kernel, so no
specific kernel-repo checkouts necessary.

Jira-Id: SP-627
Signed-off-by: Tero Tervala <tero.tervala@unikie.com>
Change-Id: Id6cacae50fb91b03bff07308b65c58a32c259a98