[Test] Upgrade buildroot to 2021.05.1
This commit is a test commit to allow using the ci
to check multiplatform results. It can be also used
by platform dvelopers to guarantee the upgrade doesn't
break their particular build.

Changelog:

[ BR2_PACKAGE_DOCKER_CONTAINERD -> BR2_PACKAGE_CONTAINERD ]
containerd is now an independent package, and it's
config reflects it in the new version of buildroot this
way.

Jira-Id: SP-608

Signed-off-by: José Pekkarinen <jose.pekkarinen@unikie.com>
Change-Id: I5bfb9553d22b9e9727fc02a5e91f385298802268