Add summary to refpolicy external modules
Modules require to have a summary to comply with the
DTD validation, if it doesn't, buildroot currently ignores
the module and continues the build, but soon it will be
checking properly the compliance, and fail the build.
This patch add the missing piece to detect the module,
build it, and pass validation.

Jira-Id: SP-610
Signed-off-by: José Pekkarinen <jose.pekkarinen@unikie.com>
Change-Id: I3195b75ff5f24ff1c9d298f0e4ae400c99c17f09