Add global patch path and test patch
This patch adds build system logic to patch buildroot
before starting using it in the different vms, by
checking the patches folder for the buildroot fixes that
are currently landed in master, but not yet in our repos.
It includes the first patch that is available in newer
versions of 2021.02, but not yet in ours 2021.02.

Jira-Id: SP-218
Signed-off-by: José Pekkarinen <jose.pekkarinen@unikie.com>
Change-Id: Ifaa678fca387a43007dc6891a9595fde8ab9e3a5