Improve check for docker image
I've been reported of a new installation where
it is visible that the old check is not enough
to find whether it is required or not to build
the image, since having git properly configured
will copy the file checked in clone step, and
it will be found in shell step. Replace it with
more accurate check.

Jira-Id: SP-218
Signed-off-by: José Pekkarinen <jose.pekkarinen@unikie.com>
Change-Id: I88c9fb1b5923e3fb1fe3c9c0eb7c9170b3d37208