Add per vm device tables
Buildroot mechanism to handle permissions in overlays
are device tables. This patch will copy the original
device table from buildroot to any of the current vms
and configure buildroot to use it instead. Currently
there is no modification on the device tables, but
it is intended to receive changes later.

Jira-Id: SP-218
Signed-off-by: José Pekkarinen <jose.pekkarinen@unikie.com>
Change-Id: I94a2e8992a90bdeeb372dc3b0459d6d4148e1d36